pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "kyma-companion"
version = "0.1.0"
description = "Kyma Companion is an AI Kyma assistant that helps you to manage your Kyma runtime"
readme = "README.md"
license = "Apache-2.0"
authors = [
  { name = "Muhammad Faizan", email = "m.faizan@sap.com" },
  { name = "Friedrich Wilken", email = "jens.wilken@sap.com" },
  { name = "Piotr Bochynski", email = "piotr.bochynski@sap.com" },
  { name = "Mansur Uralov", email = "m.uralov@sap.com" },
  { name = "Mohammad Tanweer Salah", email = "mohammad.tanweer.salah@sap.com" },
]
requires-python = "==3.13.*"
dependencies = [
  "aiohttp (>=3.13.3,<4.0.0)",
  "cryptography (>=46.0.5,<47.0.0)",
  "fastapi[standard] (>=0.135.1,<0.136.0)",
  "hdbcli (>=2.27.23,<3.0.0)",
  "kubernetes (>=35.0.0,<36.0.0)",
  "langchain (>=1.2.10,<2.0.0)",
  "langchain-core (>=1.2.17,<2.0.0)",
  "langchain-hana (>=1.0.2,<2.0.0)",
  "langfuse (>=3.14.5,<4.0.0)",
  "prometheus-client (>=0.24.1,<0.25.0)",
  "pyjwt (>=2.11.0,<3.0.0)",
  "python-decouple (>=3.8,<4.0)",
  "python-json-logger (>=4.0.0,<5.0.0)",
  "redis (>=7.2.1,<8.0.0)",
  "sap-ai-sdk-gen[all] (>=6.5.0,<7.0.0)",
  "tenacity (>=9.1.4,<10.0.0)",
  "tiktoken (>=0.12.0,<0.13.0)",
  "uvicorn (>=0.42.0,<0.43.0)"
]
[tool.poetry]
package-mode = false

[tool.poetry.group.test.dependencies]
deepeval = "^3.8.8"
fakeredis = "^2.34.1"
prettytable = "^3.17.0"
pytest = ">=9.0,<10.0"
pytest-aioresponses = "^0.3.0"
pytest-asyncio = "^1.3.0"
pytest-cov = { extras = ["all"], version = "^7.0.0" }
pytest-dotenv = "^0.5.2"
pytest-mock = "^3.15.1"
pytest-rerunfailures = ">=12.0,<17.0"
pytest-sugar = "^1.1.1"
pytest-xdist = "^3.8.0"
pyyaml = "^6.0.3"
ragas = "^0.4.3"
termcolor = "^3.3.0"
types-pyyaml = "^6.0.12.20250915"

[tool.poetry.group.dev.dependencies]
ipython = "^9.11.0"
mypy = "^1.19.1"
poethepoet = "^0.42.1"
poetry-plugin-sort = "^0.3.0"
ruff = "^0.15.4"
types-requests = "^2.32.4.20260107"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
pythonpath = ["src"]
testpaths = ["tests"]
env_files = [".env.test"]
asyncio_default_fixture_loop_scope = "function"

[tool.poe.tasks]
lint = "ruff check ."
typecheck = "mypy . --pretty"
format = "ruff format . --check"
codecheck = ["lint", "typecheck", "format"]
format-fix = "ruff format ."
lint-fix = "ruff check . --fix"
code-fix = ["format-fix", "lint-fix"]
test = "pytest -n 2 -v tests/unit"
test-local = "pytest -n auto -v tests/unit"
test-integration = "pytest -n 2 -v tests/integration -x --reruns=6 --reruns-delay=10 -r=aR"
test-integration-local = "pytest -n auto -v tests/integration -x --reruns=6 --reruns-delay=10 -r=aR"
run = "fastapi run src/main.py --port 8000"
run-local = "fastapi dev src/main.py --port 8000"
sort = "poetry sort"
pre-commit-check = ["sort", "code-fix", "codecheck", "test-local"]
